// streaming_client/src/lib/types.ts
export type IDWrapper = {
	tb: string;
	id: {
		String: string;
	};
};

export type TagOrStudio = {
	id: IDWrapper;
	name: string;
	description: string | null;
	date: string;
	slug: string;
};

export type Anime = {
	id: IDWrapper;
	mal: string;
	shinden: string | null;
	discord_role_id: string | null;
	title: string;
	description: string | null;
	anime_type: string;
	episodes: number;
	status: string;
	in_progress: boolean;
	release_time: string;
	aired: string;
	source: string;
	tags: TagOrStudio[];
	studio: TagOrStudio[];
	image: string;
	date: string;
	slug: string;
	alternative_title: string;
};
// ...
export type UserPublic = {
	id: IDWrapper;
	name: string;
	avatar: string;
	description: string;
	translator: boolean;
	proofreader: boolean;
	uploader: boolean;
	editor: boolean;
	roles: PublicRole[];
	slug: string;
};

export type LogsPublic = {
	id: IDWrapper;
	user_id: UserPublic;
	action: string;
	object: string;
	object_id: string;
	description: string | null;
	date: string;
};

export type Episodes = {
	id: IDWrapper;
	title: string | null;
	discord_id: number | null;
	translators: UserPublic[] | null;
	proofreaders: UserPublic[] | null;
	uploaders: UserPublic[] | null;
	typesetters: UserPublic[] | null;
	anime: Anime;
	episode: number;
	length: number;
	image: string | null;
	description: string | null;
	color: string;
	subtitles: string | null;
	torrent: string | null;
	video_players: string[] | null;
	discord_ping: boolean;
	webhook: boolean;
	date: string | null;
	slug: string;
};
// ...
export type PublicRole = {
	id: IDWrapper;
	name: string;
	administrative_role: boolean;
	visible: boolean;
	color: string;
	hierarchy: number;
	icon: string;
	date: string | null;
};